Create an indirection to load gihtup api client from config
There are two implementation now, the real one and the
fake agent. Which one gets used manged in configuration.
