+++ config/prod.exs
@@ -36,6 +36,8 @@ config :logger, level: :info
 
 config :faros, :book_finder, Faros.Books.GoogleFinder
 
+config :faros, :github_api_client, Faros.Github.HttpApiClient
+
 config :faros, Faros.Repo,
   adapter: Ecto.Adapters.MySQL,
     username: "${DB_USER_NAME}",
