+++ config/test.exs
@@ -11,6 +11,8 @@ config :logger, level: :warn
 
 config :faros, :book_finder, Faros.Books.FakeFinder
 
+config :faros, :github_api_client, Faros.Github.FakeApiClient
+
 config :faros, Faros.Repo,
   adapter: Ecto.Adapters.MySQL,
   username: "root",
